#5, #6: create packing npm package
