+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "pointlinejs",
-  "version": "0.1.0",
+  "version": "0.1.17",
   "description": "JavaScipt library for visualization of tree diagrams",
   "directories": {
     "example": "examples"
@@ -17,7 +17,10 @@
     "start": "npx webpack serve",
     "test": "echo \"Error: no test specified\" && exit 1",
     "doc": "typedoc --options typedoc.json",
-    "pack": "npm run build && npm pack"
+    "pack": "npm run clean && npm run build && npm run copy-css-to-es && npm run copy-css-to-lib && npm pack",
+    "copy-css-to-es": "copyfiles -u 5 src/pointlinejs/vendor/treant/styles/Treant.css dist/es/pointlinejs/vendor/treant/styles",
+    "copy-css-to-lib": "copyfiles -u 5 src/pointlinejs/vendor/treant/styles/Treant.css dist/lib/pointlinejs/vendor/treant/styles",
+    "clean": "rimraf dist/"
   },
   "keywords": [
     "pointlinejs",
@@ -41,6 +44,7 @@
     "clean-webpack-plugin": "^4.0.0",
     "copy-webpack-plugin": "^6.4.1",
     "css-loader": "^5.2.7",
+    "eve-raphael": "^0.5.0",
     "highlight.js": "^11.8.0",
     "inversify": "^6.0.1",
     "jquery.easing": "1.3",
@@ -57,6 +61,7 @@
     "@types/jquery": "^1.10.27",
     "@typescript-eslint/eslint-plugin": "^5.60.1",
     "@typescript-eslint/parser": "^5.60.1",
+    "copyfiles": "^2.4.1",
     "eslint": "^8.44.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-prettier": "^4.2.1",
@@ -66,6 +71,7 @@
     "html-webpack-plugin": "^4.5.2",
     "jquery": "^1.11.0",
     "prettier": "^2.8.0",
+    "rimraf": "^5.0.5",
     "sass": "^1.63.6",
     "sass-loader": "^13.3.2",
     "tailwindcss": "^3.3.2",
