+++ src/pointlinejs/pointlinejs.ts
@@ -11,7 +11,7 @@
  */
 
 import { ImageLoader } from './vendor/treant/ImageLoader';
-import { DI_LIST } from '@pointlinejs/InjectableList';
+import { DI_LIST } from './InjectableList';
 import { JSONconfig } from './vendor/treant/JSONConfig';
 import { NodeDB, NodeDBState } from './vendor/treant/NodeDB';
 import { ChartConfigType, Treant } from './vendor/treant/Treant';
@@ -21,6 +21,7 @@ import { TreeStore } from './vendor/treant/TreeStore';
 import { UTIL } from './vendor/treant/Util';
 import { Container } from 'inversify';
 import 'reflect-metadata';
+import { PointlineChart } from './components/pointline-chart';
 window.jQuery = window.$ = require('jquery');
 require('jquery.easing');
 
@@ -52,3 +53,5 @@ export class PointlineJS {
     return this.tree;
   }
 }
+
+export class PointlineChartTemplate extends PointlineChart { };
